tui: give the main menu its own state and the overview its own module
`screens.Context` carried the main menu's column count and visited set, so
every leaf screen was handed state only the menu uses, and the overview sat
inside the leaf-screen import cycle.
