+++ tests/unit/test_tui_curses.py
@@ -330,13 +330,13 @@ def main(window):
             ),
         ),
     )
-    context = screens.Context(
+    context = app.MainMenuContext(screens.Context(
         translate=Catalog("en"),
         disks=[("/dev/disk/by-id/virtio-target0", "40 GiB")],
         groups=load_catalog(),
         hash_password=lambda password: "$6$test$" + str(len(password)),
         timezones=("UTC",),
-    )
+    ))
     # What an operator does by opening each row and typing the disk name. The
     # per-screen behaviour is covered against the fake screen; what is under
     # test here is the whole loop on a real terminal.
